+++ src/App.jsx
@@ -1,11 +1,11 @@
 import React from "react";
 import {ThemeProvider} from "styled-components";
-import theme from "./theme";
+import theme from "./helper/theme";
 import {BrowserRouter, Routes, Route} from "react-router-dom";
 import Home from "./pages/Home";
 import Search from "./pages/Search";
-import SearchBarClickedTabProvider from "./SearchBarClickedTabProvider";
-import OptionProvider from "./OptionProvider";
+import SearchBarClickedTabProvider from "./contexts/SearchBarClickedTabProvider";
+import OptionProvider from "./contexts/OptionProvider";
 
 const App = () => {
     return (
