+++ src/components/header/modal/PriceModal.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import Modal from "../../../Modal";
+import Modal from "./Modal";
 
 const PriceModal = ({isClicked}) => {
     return <PriceModalBox isClicked={isClicked}></PriceModalBox>;
