+++ src/pages/Home.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Header from "../components/header/Header";
 import SearchBar from "../components/header/searchBar/SearchBar";
-import {size} from "../constants";
+import {size} from "../helper/constants";
 
 const Home = () => {
     return (
